Initial commit of AVD session host deployment
I need to add a note that I haven't tested this after removing all environment specific details.

I did a lot of things to make it as environment agnostic as possible, so I need to go back and ensure it does work.

Also I need to add some sort of documentation as to how this all works. :P